\E enum version for constants
This commit only introduces the use of enum instead of static final int.
The affected files are the Activity (which uses constants to send messages)
and the Service, which uses constants to read from and replyTo the Activity

In case Enum isn't needed anymore, for any reason, go back to the previous
commit, which is working without this solution, and rebase there the next
commits.

Signed-off-by: Antonio Troina <thoeni@gmail.com>